server/recentMisinformation/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
feedparser
requests
python-dotenv
pydantic